requirements.txt
Flask==1.1.4
Flask-Cors==3.0.8
flask-healthz==0.0.3
Flask-JWT-Extended==3.24.1
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.9
flask-restx==0.1.1
flatbuffers==23.1.4
gast==0.4.0
google-auth==2.16.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googletrans==3.1.0a0
grpcio==1.51.1
h11==0.9.0
h2==3.2.0
h5py==3.7.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
importlib-metadata==6.0.0
inexactsearch==1.0.2
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.2.0
jsonref==0.2
jsonschema==3.2.0
Markdown==3.4.1
MarkupSafe==2.0.1
mysql-connector-python==8.0.32
requests==2.23.0
requests-credssp==1.1.1
requests-ntlm==1.1.0
requests-oauthlib==1.3.0
Werkzeug==0.16.0

